Process all templates in CreteRepository handler.
WIP: Still need to finish ContextRepository.hbs...
