+++ GVLinQOptimizer/Programs/Utils.cs
@@ -15,4 +15,15 @@ public static class Utils
 
         return definition;
     }
+    public static async Task<string> GetResourceAsync(string fileName, CancellationToken cancellationToken)
+    {
+        var assembly = typeof(Program).Assembly;
+        var resourceName = assembly.GetManifestResourceNames().Single(n => n.EndsWith(fileName));
+        var resourceStream = assembly.GetManifestResourceStream(resourceName);
+        if (resourceStream == null)
+            throw new InvalidOperationException($"Unable to locate resource: {fileName}");
+
+        using var sr = new StreamReader(resourceStream);
+        return await sr.ReadToEndAsync(cancellationToken);
+    }
 }
