+++ GVLinQOptimizer/Templates/ContextRepository.hbs
@@ -17,6 +17,7 @@ namespace {{Namespace}}{{#newline}}
         } {{#newline}}
         {{#newline}}
         // todo: fill in methods here!! {{#newline}}
+        {{#newline}}
         #region Primary Methods{{#newline}}
         private ISQLQueryEx CreateQuery() => _factory.CreateEx(_settings.ConnectionString);{{#newline}}
         #endregion{{#newline}}
